+++ src/main/java/rip/shukaaa/keystrokes/KeyStrokeRegister.java
@@ -1,0 +1,13 @@
+package rip.shukaaa.keystrokes;
+
+import javax.swing.*;
+
+public class KeyStrokeRegister {
+		public static KeyStroke save = KeyStroke.getKeyStroke("control S"); // Save (STRG + S)
+
+		public static KeyStroke reset = KeyStroke.getKeyStroke("control shift N"); // Reset (STRG + Shift + N)
+		public static KeyStroke undo = KeyStroke.getKeyStroke("control Z"); // Undo (STRG + Z)
+		public static KeyStroke redo = KeyStroke.getKeyStroke("control Y"); // Redo (STRG + Y)
+
+		public static KeyStroke resize = KeyStroke.getKeyStroke("control R"); // Resize (STRG + R)
+}